pdf/requirements.txt
pyMuPDF
Pillow